requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain-groq
langchain_google_genai
langchain_community
pydantic
pydantic-core